moved code for dashboard into it's own separate script
