+++ server/routes/index.ts
@@ -12,6 +12,7 @@ import genericRouter from './generic'
 import prisonerSearchRoutes from '../controllers/addPrisoner/prisonerSearch'
 import addPrisonerRouter from './addPrisoner'
 import questionsRouter from './questions/router'
+import dashboard from './dashboard'
 
 export default function routes(services: Services): Router {
   const get = (path: string | string[], handler: RequestHandler) => router.get(path, asyncMiddleware(handler))
@@ -28,9 +29,9 @@ export default function routes(services: Services): Router {
 
   // view-only debug pages
   const debugRoutes = makeDebugRoutes(services)
-  get('/incidents', debugRoutes.eventList)
   get('/incidents/:id', debugRoutes.eventDetails)
   get('/reports/:id', debugRoutes.reportDetails)
+  router.use('/incidents', dashboard(services))
 
   // report pages
   router.use('/create-report', createReportRouter)
